Modifica de int a long long, las iteraciones, para trabajar con números más grandes

+++ Actividades/Monte Carlo y Dartboard/MonteCarloDart.c
@@ -4,16 +4,16 @@
 #include <math.h>
 
 //Funciones prototipo
-int monte_carlo(int);
+int monte_carlo(long long);
 
 
 
 
 int main(int argc, char *argv[]){
-    int iteraciones = (int)atoi(argv[1]);
+    long long iteraciones = atoll(argv[1]);
     float pi;
     int i, p_circulo=0;
-
+    
     //Define la semilla usando el reloj del sistema
     srand(time(NULL));
     clock_t inicio = clock();
@@ -27,22 +27,22 @@ int main(int argc, char *argv[]){
     return 0;
 }
 
-int monte_carlo(int iteraciones){
+int monte_carlo(long long iteraciones){
     float x, y, d;
     int p_circulo=0;
 
-    for(int i=0; i<iteraciones; i++){
+    for(long long i=0; i<iteraciones; i++){
         //Genera una coordenada (x,y)
         x = (float)rand()/(float)(RAND_MAX/2);
         x = x - 1;
         y = (float)rand()/(float)(RAND_MAX/2);
         y = y - 1;
-
         //Calcula si el dardo cae en el circulo
         d = (x*x) + (y*y);
         if(d<=1){
             p_circulo++;
         }
+        
     }
     return p_circulo;
 }
